% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amorem-package.R
\docType{package}
\name{amorem-package}
\alias{amorem}
\alias{amorem-package}
\title{amorem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utilities for simulating and fittin relational event models, including helpers to generate dynamic event sequences and covariate processes for sender and receiver sets. The endogenous-effect and case-control estimation machinery follows Juozaitiene and Wit (2024) \doi{10.1093/jrsssa/qnae132}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://franciscorichter.github.io/amorem/}
  \item \url{https://github.com/franciscorichter/amorem}
}

}
\author{
\strong{Maintainer}: Francisco Richter \email{richtf@usi.ch}

Authors:
\itemize{
  \item Martina Boschi
  \item Ernst C. Wit
  \item Melania Lembo
}

}
\keyword{internal}
