% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{email_eu_core}
\alias{email_eu_core}
\title{Email-Eu-Core temporal (Department 3 subset)}
\format{
A data frame with 12,216 rows and 3 columns:
\describe{
\item{time}{Days since the first email in the recording window.}
\item{sender}{Character employee id (anonymised).}
\item{receiver}{Character employee id (anonymised).}
}
}
\source{
Paranjape, A., Benson, A.R., Leskovec, J. (2017). Motifs in
temporal networks. \emph{WSDM '17}, 601–610.
\doi{10.1145/3018661.3018731}. Distributed via SNAP:
\url{https://snap.stanford.edu/data/email-Eu-core-temporal.html}.
}
\usage{
email_eu_core
}
\description{
Internal emails between members of a single department of a large
European research institution over ~803 days. The dataset has been
filtered to remove self-loops. Sourced from the SNAP repository as
the Department 3 slice of the email-Eu-core-temporal dataset.
}
\keyword{datasets}
