% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperedge_log.R
\name{hyperedge_sizes}
\alias{hyperedge_sizes}
\title{Cardinality columns for a hyperedge event log}
\usage{
hyperedge_sizes(hyperedge_log)
}
\arguments{
\item{hyperedge_log}{A hyperedge log.}
}
\value{
The hyperedge log with two added integer columns.
}
\description{
Adds two integer columns to a hyperedge log: \code{size_I} (the number
of senders) and \code{size_J} (the number of receivers). Convenient
shortcut for filtering / case-control sampling matched on
cardinality (see Boschi et al. 2025, Section 3.3).
}
