% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_nn_uncertainty.R
\name{plot.nn_uncertainty}
\alias{plot.nn_uncertainty}
\title{Plot partial-dependence uncertainty bands}
\usage{
\method{plot}{nn_uncertainty}(x, ...)
}
\arguments{
\item{x}{An \code{\link[=nn_uncertainty]{nn_uncertainty()}} object.}

\item{...}{Passed to the underlying \code{plot()}.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot partial-dependence uncertainty bands
}
