% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amore-package.R
\docType{package}
\name{amore-package}
\alias{amore}
\alias{amore-package}
\title{amore: Augmented Modelling of Relational Events}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utilities for simulating and prototyping relational event models, including helpers to generate dynamic event sequences and covariate processes for sender and receiver sets. The endogenous-effect and case-control estimation machinery follows Juozaitiene and Wit (2024) \doi{10.1093/jrsssa/qnae132}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://franciscorichter.github.io/amore/}
  \item \url{https://github.com/franciscorichter/amore}
}

}
\author{
\strong{Maintainer}: Francisco Richter \email{richtf@usi.ch}

}
\keyword{internal}
